python/requirements.txt
# Python dependencies for Praxus backend
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
pydantic==2.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
# Add more as needed